Removed unnecessary cast and brought some documentation up-to-date.

+++ zee/src/completion.c
@@ -89,7 +89,7 @@ static size_t common_prefix_length(astr as, astr bs) {
   size_t len = min(astr_len(as), astr_len(bs));
   for (size_t i = 0; i < len; i++)
     if (*astr_char(as, (ptrdiff_t)i) != *astr_char(bs, (ptrdiff_t)i))
-      return (size_t)i;
+      return i;
   return len;
 }
